Move form builder "button component" (#2072)
* update path
