+++ components/form-builder/app/SetResponseDelivery.tsx
@@ -7,7 +7,8 @@ import { useSession } from "next-auth/react";
 import { useRefresh } from "@lib/hooks";
 import { isValidGovEmail } from "@lib/validation";
 import { ResponseEmail } from "./ResponseEmail";
-import { Radio, Button } from "./shared";
+import { Radio } from "./shared";
+import { Button } from "@components/globals";
 import { useTemplateApi } from "../hooks";
 import { useTemplateStore } from "../store";
 import { completeEmailAddressRegex } from "../util";
