+++ components/form-builder/app/edit/elements/SubOption.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "next-i18next";
 import debounce from "lodash.debounce";
 
 import { Close } from "../../../icons";
-import { Button } from "../../shared/Button";
+import { Button } from "@components/globals";
 import { Input } from "../../shared/Input";
 import { useTemplateStore } from "../../../store/useTemplateStore";
 import { Language } from "../../../types";
