+++ components/form-builder/app/navigation/ShareDropdown.tsx
@@ -9,7 +9,7 @@ import { ChevronDown, ChevronRight, ShareIcon, LinkIcon } from "../../icons";
 import { useTemplateStore } from "../../store/useTemplateStore";
 import { ShareModal } from "../ShareModal";
 import { LinksSubMenu } from "./LinksSubMenu";
-import { Button } from "../shared/Button";
+import { Button } from "@components/globals";
 import { ShareModalUnauthenticated } from "..";
 
 export const ShareDropdown = () => {
