+++ components/globals/Buttons/index.ts
@@ -1,0 +1,2 @@
+export { themes } from "./themes";
+export { Button } from "./Button";
