+++ components/globals/index.ts
@@ -1,0 +1,1 @@
+export { themes, Button } from "./Buttons";
